refactor(cli): broke up cli code
