+++ packages/xinity-cli/src/lib/dashboard-stubs.ts
@@ -139,7 +139,7 @@ const sourceOverrides: Record<string, string> = {
 
 /** Convert an object-form stub to ESM source string for Bun.build's onLoad. */
 function toJsSource(specifier: string, exports: Record<string, unknown>): string {
-  if (specifier in sourceOverrides) return sourceOverrides[specifier];
+  if (specifier in sourceOverrides) return sourceOverrides[specifier]!;
 
   const lines: string[] = [];
   for (const [name, value] of Object.entries(exports)) {
